# Initial refactoring of edge storage