--- a/vpr/src/draw/draw.h
+++ b/vpr/src/draw/draw.h
@@ -32,7 +32,7 @@ void free_draw_structs();
 #ifndef NO_GRAPHICS
 
 void draw_get_rr_pin_coords(int inode, float* xcen, float* ycen);
-void draw_get_rr_pin_coords(const t_rr_node* node, float* xcen, float* ycen);
+void draw_get_rr_pin_coords(const t_rr_node node, float* xcen, float* ycen);
 
 void draw_triangle_along_line(ezgl::renderer* g, ezgl::point2d start, ezgl::point2d end, float relative_position = 1., float arrow_size = DEFAULT_ARROW_SIZE);
 void draw_triangle_along_line(ezgl::renderer* g, ezgl::point2d loc, ezgl::point2d start, ezgl::point2d end, float arrow_size = DEFAULT_ARROW_SIZE);