--- a/vpr/src/power/power.cpp
+++ b/vpr/src/power/power.cpp
@@ -815,19 +815,19 @@ static void power_usage_routing(t_power_usage* power_usage,
         t_trace* trace;
 
         for (trace = route_ctx.trace[net_id].head; trace != nullptr; trace = trace->next) {
-            auto node = &device_ctx.rr_nodes[trace->index];
+            auto node = device_ctx.rr_nodes[trace->index];
             t_rr_node_power* node_power = &rr_node_power[trace->index];
 
             if (node_power->visited) {
                 continue;
             }
 
-            for (t_edge_size edge_idx = 0; edge_idx < node->num_edges(); edge_idx++) {
-                if (node->edge_sink_node(edge_idx) != OPEN) {
-                    auto next_node = &device_ctx.rr_nodes[node->edge_sink_node(edge_idx)];
-                    t_rr_node_power* next_node_power = &rr_node_power[node->edge_sink_node(edge_idx)];
+            for (t_edge_size edge_idx = 0; edge_idx < node.num_edges(); edge_idx++) {
+                if (node.edge_sink_node(edge_idx) != OPEN) {
+                    auto next_node = device_ctx.rr_nodes[node.edge_sink_node(edge_idx)];
+                    t_rr_node_power* next_node_power = &rr_node_power[node.edge_sink_node(edge_idx)];
 
-                    switch (next_node->type()) {
+                    switch (next_node.type()) {
                         case CHANX:
                         case CHANY:
                         case IPIN:
@@ -837,9 +837,9 @@ static void power_usage_routing(t_power_usage* power_usage,
                             next_node_power->in_dens[next_node_power->num_inputs] = clb_net_density(node_power->net_num);
                             next_node_power->in_prob[next_node_power->num_inputs] = clb_net_prob(node_power->net_num);
                             next_node_power->num_inputs++;
-                            if (next_node_power->num_inputs > next_node->fan_in()) {
+                            if (next_node_power->num_inputs > next_node.fan_in()) {
                                 VTR_LOG("%d %d\n", next_node_power->num_inputs,
-                                        next_node->fan_in());
+                                        next_node.fan_in());
                                 fflush(nullptr);
                                 VTR_ASSERT(0);
                             }
@@ -857,7 +857,7 @@ static void power_usage_routing(t_power_usage* power_usage,
     /* Calculate power of all routing entities */
     for (size_t rr_node_idx = 0; rr_node_idx < device_ctx.rr_nodes.size(); rr_node_idx++) {
         t_power_usage sub_power_usage;
-        auto node = &device_ctx.rr_nodes[rr_node_idx];
+        auto node = device_ctx.rr_nodes[rr_node_idx];
         t_rr_node_power* node_power = &rr_node_power[rr_node_idx];
         float C_wire;
         float buffer_size;
@@ -866,7 +866,7 @@ static void power_usage_routing(t_power_usage* power_usage,
         //float C_per_seg_split;
         int wire_length;
 
-        switch (node->type()) {
+        switch (node.type()) {
             case SOURCE:
             case SINK:
             case OPIN:
@@ -877,13 +877,13 @@ static void power_usage_routing(t_power_usage* power_usage,
                  *  - Driver (accounted for at end of CHANX/Y - see below)
                  *  - Multiplexor */
 
-                if (node->fan_in()) {
+                if (node.fan_in()) {
                     VTR_ASSERT(node_power->in_dens);
                     VTR_ASSERT(node_power->in_prob);
 
                     /* Multiplexor */
                     power_usage_mux_multilevel(&sub_power_usage,
-                                               power_get_mux_arch(node->fan_in(),
+                                               power_get_mux_arch(node.fan_in(),
                                                                   power_ctx.arch->mux_transistor_size),
                                                node_power->in_prob, node_power->in_dens,
                                                node_power->selected_input, true,
@@ -904,19 +904,19 @@ static void power_usage_routing(t_power_usage* power_usage,
                 VTR_ASSERT(node_power->in_prob);
 
                 wire_length = 0;
-                if (node->type() == CHANX) {
-                    wire_length = node->xhigh() - node->xlow() + 1;
-                } else if (node->type() == CHANY) {
-                    wire_length = node->yhigh() - node->ylow() + 1;
+                if (node.type() == CHANX) {
+                    wire_length = node.xhigh() - node.xlow() + 1;
+                } else if (node.type() == CHANY) {
+                    wire_length = node.yhigh() - node.ylow() + 1;
                 }
                 C_wire = wire_length
-                         * segment_inf[device_ctx.rr_indexed_data[node->cost_index()].seg_index].Cmetal;
+                         * segment_inf[device_ctx.rr_indexed_data[node.cost_index()].seg_index].Cmetal;
                 //(double)power_ctx.commonly_used->tile_length);
-                VTR_ASSERT(node_power->selected_input < node->fan_in());
+                VTR_ASSERT(node_power->selected_input < node.fan_in());
 
                 /* Multiplexor */
                 power_usage_mux_multilevel(&sub_power_usage,
-                                           power_get_mux_arch(node->fan_in(),
+                                           power_get_mux_arch(node.fan_in(),
                                                               power_ctx.arch->mux_transistor_size),
                                            node_power->in_prob, node_power->in_dens,
                                            node_power->selected_input, true, power_ctx.solution_inf.T_crit);
@@ -979,10 +979,10 @@ static void power_usage_routing(t_power_usage* power_usage,
                 /* Determine types of switches that this wire drives */
                 connectionbox_fanout = 0;
                 switchbox_fanout = 0;
-                for (t_edge_size iedge = 0; iedge < node->num_edges(); iedge++) {
-                    if (node->edge_switch(iedge) == routing_arch->wire_to_rr_ipin_switch) {
+                for (t_edge_size iedge = 0; iedge < node.num_edges(); iedge++) {
+                    if (node.edge_switch(iedge) == routing_arch->wire_to_rr_ipin_switch) {
                         connectionbox_fanout++;
-                    } else if (node->edge_switch(iedge) == routing_arch->delayless_switch) {
+                    } else if (node.edge_switch(iedge) == routing_arch->delayless_switch) {
                         /* Do nothing */
                     } else {
                         switchbox_fanout++;
@@ -1205,37 +1205,37 @@ void power_routing_init(const t_det_routing_arch* routing_arch) {
     for (size_t rr_node_idx = 0; rr_node_idx < device_ctx.rr_nodes.size(); rr_node_idx++) {
         int fanout_to_IPIN = 0;
         int fanout_to_seg = 0;
-        auto node = &device_ctx.rr_nodes[rr_node_idx];
+        auto node = device_ctx.rr_nodes[rr_node_idx];
         t_rr_node_power* node_power = &rr_node_power[rr_node_idx];
 
-        switch (node->type()) {
+        switch (node.type()) {
             case IPIN:
                 max_IPIN_fanin = std::max(max_IPIN_fanin,
-                                          static_cast<int>(node->fan_in()));
-                max_fanin = std::max(max_fanin, static_cast<int>(node->fan_in()));
+                                          static_cast<int>(node.fan_in()));
+                max_fanin = std::max(max_fanin, static_cast<int>(node.fan_in()));
 
-                node_power->in_dens = (float*)vtr::calloc(node->fan_in(),
+                node_power->in_dens = (float*)vtr::calloc(node.fan_in(),
                                                           sizeof(float));
-                node_power->in_prob = (float*)vtr::calloc(node->fan_in(),
+                node_power->in_prob = (float*)vtr::calloc(node.fan_in(),
                                                           sizeof(float));
                 break;
             case CHANX:
             case CHANY:
-                for (t_edge_size iedge = 0; iedge < node->num_edges(); iedge++) {
-                    if (node->edge_switch(iedge) == routing_arch->wire_to_rr_ipin_switch) {
+                for (t_edge_size iedge = 0; iedge < node.num_edges(); iedge++) {
+                    if (node.edge_switch(iedge) == routing_arch->wire_to_rr_ipin_switch) {
                         fanout_to_IPIN++;
-                    } else if (node->edge_switch(iedge) != routing_arch->delayless_switch) {
+                    } else if (node.edge_switch(iedge) != routing_arch->delayless_switch) {
                         fanout_to_seg++;
                     }
                 }
                 max_seg_to_IPIN_fanout = std::max(max_seg_to_IPIN_fanout,
                                                   fanout_to_IPIN);
                 max_seg_to_seg_fanout = std::max(max_seg_to_seg_fanout, fanout_to_seg);
-                max_fanin = std::max(max_fanin, static_cast<int>(node->fan_in()));
+                max_fanin = std::max(max_fanin, static_cast<int>(node.fan_in()));
 
-                node_power->in_dens = (float*)vtr::calloc(node->fan_in(),
+                node_power->in_dens = (float*)vtr::calloc(node.fan_in(),
                                                           sizeof(float));
-                node_power->in_prob = (float*)vtr::calloc(node->fan_in(),
+                node_power->in_prob = (float*)vtr::calloc(node.fan_in(),
                                                           sizeof(float));
                 break;
             default:
@@ -1254,14 +1254,14 @@ void power_routing_init(const t_det_routing_arch* routing_arch) {
 
     /* Populate driver switch type */
     for (size_t rr_node_idx = 0; rr_node_idx < device_ctx.rr_nodes.size(); rr_node_idx++) {
-        auto node = &device_ctx.rr_nodes[rr_node_idx];
+        auto node = device_ctx.rr_nodes[rr_node_idx];
 
-        for (t_edge_size edge_idx = 0; edge_idx < node->num_edges(); edge_idx++) {
-            if (node->edge_sink_node(edge_idx) != OPEN) {
-                if (rr_node_power[node->edge_sink_node(edge_idx)].driver_switch_type == OPEN) {
-                    rr_node_power[node->edge_sink_node(edge_idx)].driver_switch_type = node->edge_switch(edge_idx);
+        for (t_edge_size edge_idx = 0; edge_idx < node.num_edges(); edge_idx++) {
+            if (node.edge_sink_node(edge_idx) != OPEN) {
+                if (rr_node_power[node.edge_sink_node(edge_idx)].driver_switch_type == OPEN) {
+                    rr_node_power[node.edge_sink_node(edge_idx)].driver_switch_type = node.edge_switch(edge_idx);
                 } else {
-                    VTR_ASSERT(rr_node_power[node->edge_sink_node(edge_idx)].driver_switch_type == node->edge_switch(edge_idx));
+                    VTR_ASSERT(rr_node_power[node.edge_sink_node(edge_idx)].driver_switch_type == node.edge_switch(edge_idx));
                 }
             }
         }
@@ -1270,13 +1270,13 @@ void power_routing_init(const t_det_routing_arch* routing_arch) {
     /* Find Max Fanout of Routing Buffer	 */
     t_edge_size max_seg_fanout = 0;
     for (size_t rr_node_idx = 0; rr_node_idx < device_ctx.rr_nodes.size(); rr_node_idx++) {
-        auto node = &device_ctx.rr_nodes[rr_node_idx];
+        auto node = device_ctx.rr_nodes[rr_node_idx];
 
-        switch (node->type()) {
+        switch (node.type()) {
             case CHANX:
             case CHANY:
-                if (node->num_edges() > max_seg_fanout) {
-                    max_seg_fanout = node->num_edges();
+                if (node.num_edges() > max_seg_fanout) {
+                    max_seg_fanout = node.num_edges();
                 }
                 break;
             default:
@@ -1358,14 +1358,14 @@ bool power_uninit() {
     bool error = false;
 
     for (size_t rr_node_idx = 0; rr_node_idx < device_ctx.rr_nodes.size(); rr_node_idx++) {
-        auto node = &device_ctx.rr_nodes[rr_node_idx];
+        auto node = device_ctx.rr_nodes[rr_node_idx];
         t_rr_node_power* node_power = &rr_node_power[rr_node_idx];
 
-        switch (node->type()) {
+        switch (node.type()) {
             case CHANX:
             case CHANY:
             case IPIN:
-                if (node->fan_in()) {
+                if (node.fan_in()) {
                     free(node_power->in_dens);
                     free(node_power->in_prob);
                 }