#### Description

This changes edge storage from an allocation array of struct per node to
struct of array for all edge data.

Several algorithms over edges that were previous per node per edge, but
were actually just iteration over edges are now part of rr_node_storage.

This PR is built on top #1084 .

#### Related Issue

#1079 
#1081 
#1084 

#### Motivation and Context

This changes the number of heap allocations on preallocated array (e.g. loading rr graph from file) from `Nnodes` to `3`.

This changes reduces the number of heap allocations during rr graph building from `Nnodes` to `3 * ceil(log2(Nedges) - log2(10*Nodes))`.

When the number of edges is known the max memory usage for edges is now `Nedge * (2*sizeof(int)+sizeof(short))`.  When the number of edges is not known, the max memory usage for edges is now:

```
min(10*Nnode, Nedge) * (2*sizeof(int)+sizeof(short)) + sizeof(int)* (Nedge/2)
```

#### How Has This Been Tested?

- [ ] Travis CI is green
- [ ] Nightly and weekly QoR metrics are acceptable

#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (change which fixes an issue)
- [x] New feature (change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

#### Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My change requires a change to the documentation
- [x] I have updated the documentation accordingly
- [ ] I have added tests to cover my changes
- [x] All new and existing tests passed