@HackerFoo 2020-01-27
There are many uses of `first` and `second` seem to be used as placeholder names, like `x` and `y`.