@litghost 2020-01-28
> Sanity check with `gsm_switch_stratixiv_arch_timing.blif` (from weekly titan tests)
> 
> Before:
> 
> ```
> ## Build routing resource graph took 212.56 seconds (max_rss 7950.6 MiB, delta_rss +1698.9 MiB)
>   RR Graph Nodes: 20014774
>   RR Graph Edges: 159049032
> # Create Device took 214.84 seconds (max_rss 7950.6 MiB, delta_rss +1698.9 MiB)
> ```
> 
> After:
> 
> ```
> ## Build routing resource graph took 159.88 seconds (max_rss 9165.3 MiB, delta_rss +2842.8 MiB)
>   RR Graph Nodes: 20014774
>   RR Graph Edges: 159049032
> # Create Device took 161.22 seconds (max_rss 9165.3 MiB, delta_rss +2842.8 MiB)
> ```
> 
> This jump in ~1.2 GiB is basically the cost of storing the source rr node explicitly. I'll noddle over alternative strategies.

So in an unexpected turn, while the max_rss after create device **is higher** than before, because the heap as not threshed, the final max_rss is much better:

Before:
```
The entire flow of VPR took 13322.02 seconds (max_rss 16152.4 MiB)
```

After:
```
The entire flow of VPR took 7068.79 seconds (max_rss 10713.4 MiB)
```

I'll need to do a full `vtr_reg_weekly` to confirm the results.