@litghost 2020-02-03
I think I've identified a way to regain some lost CPU performance.  I've started a vtr_reg_weekly QoR run, but it won't return results for a day or two.

I'll push a new branch today with a rebase onto of master, along with fixes to review comments.