@litghost 2020-02-03
I believe all feedback has been addressed.  I've kicked off a QoR run, results will be ready in a couple days.  Preliminary results show the CPU cost is gone, with the 4% memory increase remaining.  #1096 resolves the memory increase, and changes it to a 35-60% memory reduction.